+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## Upcoming
+
+- There was a timing issue in that if 'fetchMediaItemsById' (steps/source-nodes/fetch-nodes/fetch-referenced-media-items.js)
+  is called a 2nd time before the first batch completes the call to pushPromiseOntoRetryQueue overwrites keys from the first batch as the key is only an index. Then the first instance item never resolves and they all timeout.
+
 ## 1.7.5
 
 ### Bug Fixes